Add `retry` to the `get_deployments` function
I've been seeing a few of these errors from get_deployments:
"the connection is closed for writing, which means that you cannot issue any more requests on the connection but you can expect responses to still be delivered for part of the requests that are in flight. If a connection is closed for writing, it usually means that you got a :server_closed_request error already."
I have a feeling this is a k8s api server load issue as we have a lot of application waking up at the same time. As a short term fix I'm implementing retries to simply try fetching the deployment info again. Long term I'd like to add some jitter to the sleep schedules so that it doesn't create a big burst of traffic when everything starts to wake or sleep.